Leave more of the page space for actual content.
The current layout leaves many pixels between the left navigation
bar and the content of the body. This is a static value.

In this change, removed the static spacing, left the padding of
the parent object, which seems to be more than enough.

Look at the difference in the two screenshots:

https://ibb.co/jDz3Ha
https://ibb.co/bUFzqv

Note: I tested this locally and with the default phone simulator
in chromium. Scrolled through the pages, and they look ok, including
tables.